LAMBDA_LABS/human-rights-first-ds-e/project/requirements.txt
fastapi==0.60.1
pandas==1.1.0
plotly==4.9.0
uvicorn==0.11.8
praw==7.1.0
python-dotenv==0.14.0
beautifulsoup4==4.9.1
scikit-learn==0.23.2
spacy==2.3.2
newspaper3k==0.2.8
fastapi-utils==0.2.1
psycopg2==2.8